mc5g-metainfo-website-backend/requirements.txt
motor==3.3.2
oauthlib==3.2.2
PyYAML==6.0.1
quart==0.19.4
requests==2.31.0
ujenkins==0.10.2
Werkzeug==3.0.1